requirements.txt
numpy @ file:///C:/b/abs_5ct9ex77k9/croot/numpy_and_numpy_base_1668593740598/work
opencv-python==4.6.0.66
Pillow==9.2.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
csv==1.0
